// backend/src/generated/api.ts
export const CreateReportBody = zod.object({
  treeId: zod.number().nullish(),
  treeCode: zod.string().nullish(),
  reportType: zod.enum([
    "plantation",
    "cutting",
    "survival_check",
    "illegal_cutting",
  ]),
  reportedBy: zod.string(),
  description: zod.string(),
  photoUrl: zod.string().nullish(),
  latitude: zod.number().nullish(),
  longitude: zod.number().nullish(),
  state: zod.string(),
  district: zod.string(),
});